+++ src/main/java/com/evacipated/cardcrawl/modthespire/patcher/NonStaticPatchMethodException.java
@@ -1,0 +1,11 @@
+package com.evacipated.cardcrawl.modthespire.patcher;
+
+import javassist.CtMethod;
+
+public class NonStaticPatchMethodException extends Exception
+{
+    public NonStaticPatchMethodException(CtMethod patchMethod)
+    {
+        super(patchMethod.getLongName());
+    }
+}