+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Changelog ##
 #### dev ####
+* Add clear error if patch method is non-static
 
 #### v3.28.0 ####
 * Bump gson from 2.8.2 to 2.8.9